add spec for fdja_[value_]at()

+++ src/djan.h
@@ -196,7 +196,7 @@ size_t fdja_size(fdja_value *v);
  */
 int fdja_cmp(fdja_value *a, fdja_value *b);
 
-fdja_value *fdja_value_at(fdja_value *v, long n);
+fdja_value *fdja_value_at(fdja_value *v, long long n);
 #define fdja_at(v, n) fdja_value_at(v, n)
 
 fdja_value *fdja_vlookup(fdja_value *v, const char *path, va_list ap);
